lara/col/jump: separate jumping routines
This loosely groups Lara's jumping collision routines, and those
related to freefall.
