+++ src/libtrx/game/lara/col.c
@@ -39,19 +39,10 @@ static void M_Stop(ITEM *item, COLL_INFO *coll);
 static void M_FastBack(ITEM *item, COLL_INFO *coll);
 static void M_Turn(ITEM *item, COLL_INFO *coll);
 static void M_Death(ITEM *item, COLL_INFO *coll);
-static void M_FastFall(ITEM *item, COLL_INFO *coll);
-static void M_Reach(ITEM *item, COLL_INFO *coll);
 static void M_Splat(ITEM *item, COLL_INFO *coll);
-static void M_Compress(ITEM *item, COLL_INFO *coll);
 static void M_Slide(ITEM *item, COLL_INFO *coll);
-static void M_ForwardJump(ITEM *item, COLL_INFO *coll);
-static void M_UpJump(ITEM *item, COLL_INFO *coll);
-static void M_SideBackJump(ITEM *item, COLL_INFO *coll);
-static void M_FallBack(ITEM *item, COLL_INFO *coll);
 static void M_Roll(ITEM *item, COLL_INFO *coll);
 static void M_RollContinue(ITEM *item, COLL_INFO *coll);
-static void M_SwanDive(ITEM *item, COLL_INFO *coll);
-static void M_FastDive(ITEM *item, COLL_INFO *coll);
 
 static void (*m_CollisionRoutines[LS_NUMBER_OF])(
     ITEM *item, COLL_INFO *coll) = {
@@ -59,17 +50,13 @@ static void (*m_CollisionRoutines[LS_NUMBER_OF])(
     [LS_WALK]         = M_Walk,
     [LS_RUN]          = M_Run,
     [LS_STOP]         = M_Stop,
-    [LS_JUMP_FORWARD] = M_ForwardJump,
     [LS_POSE]         = M_Stop,
     [LS_FAST_BACK]    = M_FastBack,
     [LS_TURN_RIGHT]   = M_Turn,
     [LS_TURN_LEFT]    = M_Turn,
     [LS_DEATH]        = M_Death,
-    [LS_FAST_FALL]    = M_FastFall,
-    [LS_REACH]        = M_Reach,
     [LS_SPLAT]        = M_Splat,
     [LS_LAND]         = M_Stop,
-    [LS_COMPRESS]     = M_Compress,
     [LS_WALK_BACK]    = M_WalkBack,
     [LS_PULL_UP]      = M_Default,
     [LS_FAST_TURN]    = M_Stop,
@@ -77,11 +64,6 @@ static void (*m_CollisionRoutines[LS_NUMBER_OF])(
     [LS_STEP_LEFT]    = M_SideStep,
     [LS_ROLL_CONT]    = M_RollContinue,
     [LS_SLIDE]        = M_Slide,
-    [LS_JUMP_BACK]    = M_SideBackJump,
-    [LS_JUMP_RIGHT]   = M_SideBackJump,
-    [LS_JUMP_LEFT]    = M_SideBackJump,
-    [LS_JUMP_UP]      = M_UpJump,
-    [LS_FALL_BACK]    = M_FallBack,
     [LS_SLIDE_BACK]   = M_Slide,
     [LS_PUSH_BLOCK]   = M_Default,
     [LS_PULL_BLOCK]   = M_Default,
@@ -94,8 +76,6 @@ static void (*m_CollisionRoutines[LS_NUMBER_OF])(
     [LS_ROLL]         = M_Roll,
     [LS_USE_MIDAS]    = M_Default,
     [LS_DIE_MIDAS]    = M_Default,
-    [LS_SWAN_DIVE]    = M_SwanDive,
-    [LS_FAST_DIVE]    = M_FastDive,
     [LS_GYMNAST]      = M_Default,
     [LS_WATER_OUT]    = M_Default,
 #if TR_VERSION == 1
@@ -441,92 +421,10 @@ static void M_Death(ITEM *const item, COLL_INFO *const coll)
     lara->air = -1;
 }
 
-static void M_FastFall(ITEM *const item, COLL_INFO *const coll)
-{
-    item->gravity = true;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    Lara_SlideEdgeJump(item, coll);
-    if (coll->side_mid.floor > 0) {
-        return;
-    }
-
-    if (Lara_LandedBad(item, coll)) {
-        item->goal_anim_state = LS_DEATH;
-    } else {
-        item->goal_anim_state = LS_STOP;
-        item->current_anim_state = LS_STOP;
-        Item_SwitchToAnim(item, LA_FREEFALL_LAND, 0);
-    }
-
-    Sound_StopEffect(SFX_LARA_FALL);
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
-}
-
-static void M_Reach(ITEM *const item, COLL_INFO *const coll)
-{
-    item->gravity = true;
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    lara->move_angle = item->rot.y;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = 0;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    if (Lara_TestHangJump(item, coll)) {
-        return;
-    }
-
-    Lara_SlideEdgeJump(item, coll);
-    if (item->fall_speed <= 0 || coll->side_mid.floor > 0) {
-        return;
-    }
-
-    if (Lara_LandedBad(item, coll)) {
-        item->goal_anim_state = LS_DEATH;
-    } else {
-        item->goal_anim_state = LS_STOP;
-    }
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
-}
-
 static void M_Splat(ITEM *const item, COLL_INFO *const coll)
 {
     M_Default(item, coll);
     Lara_ShiftCol(coll);
-#if TR_VERSION >= 2
-    if (coll->side_mid.floor > -STEP_L && coll->side_mid.floor < STEP_L) {
-        item->pos.y += coll->side_mid.floor;
-    }
-#endif
-}
-
-static void M_Compress(ITEM *const item, COLL_INFO *const coll)
-{
-    item->gravity = false;
-    item->fall_speed = 0;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = NO_BAD_NEG;
-    coll->bad_ceiling = 0;
-
-    Lara_GetCollisionInfo(item, coll);
-
-    if (coll->side_mid.ceiling > -100) {
-        Item_SwitchToAnim(item, LA_STAND_STILL, 0);
-        item->goal_anim_state = LS_STOP;
-        item->current_anim_state = LS_STOP;
-        item->gravity = false;
-        item->speed = 0;
-        item->fall_speed = 0;
-        item->pos = coll->old;
-    }
 #if TR_VERSION >= 2
     if (coll->side_mid.floor > -STEP_L && coll->side_mid.floor < STEP_L) {
         item->pos.y += coll->side_mid.floor;
@@ -542,171 +440,6 @@ static void M_Slide(ITEM *const item, COLL_INFO *const coll)
         lara->move_angle += DEG_180;
     }
     Lara_SlideSlope(item, coll);
-}
-
-static void M_ForwardJump(ITEM *const item, COLL_INFO *const coll)
-{
-#if TR_VERSION == 1
-    // TODO: TR1's wall bug actually stems from Lara_DeflectEdgeJump, see about
-    // fixing it there.
-    const bool backward_momentum = false;
-    const bool fix_wall_bug = g_Config.gameplay.fix_wall_jump_glitch;
-#else
-    const bool backward_momentum = item->speed < 0;
-    const bool fix_wall_bug = false;
-#endif
-
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    if (backward_momentum) {
-        lara->move_angle = item->rot.y + DEG_180;
-    } else {
-        lara->move_angle = item->rot.y;
-    }
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    Lara_DeflectEdgeJump(item, coll);
-    if (backward_momentum) {
-        lara->move_angle = item->rot.y;
-    }
-
-    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
-        return;
-    }
-
-    if (Lara_LandedBad(item, coll)) {
-        item->goal_anim_state = LS_DEATH;
-    } else if (
-        lara->water_status != LWS_WADE && g_Input.forward && !g_Input.slow) {
-        item->goal_anim_state = LS_RUN;
-    } else {
-        item->goal_anim_state = LS_STOP;
-    }
-
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
-    item->speed = 0;
-    if (!fix_wall_bug) {
-        Lara_Animate(item);
-    }
-}
-
-static void M_UpJump(ITEM *const item, COLL_INFO *const coll)
-{
-#if TR_VERSION == 1
-    const bool enable_lean_jumping = g_Config.gameplay.enable_lean_jumping;
-#else
-    const bool enable_lean_jumping = true;
-#endif
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    lara->move_angle = item->rot.y;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-    coll->facing = lara->move_angle;
-    if (enable_lean_jumping && item->speed < 0) {
-        coll->facing += DEG_180;
-    }
-
-    Collide_GetCollisionInfo(
-        coll, item->pos.x, item->pos.y, item->pos.z, item->room_num, 870);
-    if (Lara_TestHangJumpUp(item, coll)) {
-        return;
-    }
-
-    Lara_SlideEdgeJump(item, coll);
-    if (enable_lean_jumping) {
-        if (coll->coll_type != COLL_NONE) {
-            item->speed = item->speed > 0 ? 2 : -2;
-        } else if (item->fall_speed < -70) {
-            if (g_Input.forward && item->speed < 5) {
-                item->speed++;
-            } else if (g_Input.back && item->speed > -5) {
-                item->speed -= 2;
-            }
-        }
-    }
-
-    if (item->fall_speed <= 0 || coll->side_mid.floor > 0) {
-        return;
-    }
-
-    if (Lara_LandedBad(item, coll)) {
-        item->goal_anim_state = LS_DEATH;
-    } else {
-        item->goal_anim_state = LS_STOP;
-    }
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
-}
-
-static void M_SideBackJump(ITEM *const item, COLL_INFO *const coll)
-{
-    int32_t angle = 0;
-    switch (item->current_anim_state) {
-    case LS_JUMP_BACK:
-        angle = DEG_180;
-        break;
-    case LS_JUMP_RIGHT:
-        angle = DEG_90;
-        break;
-    case LS_JUMP_LEFT:
-        angle = -DEG_90;
-        break;
-    default:
-        return;
-    }
-
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    lara->move_angle = item->rot.y + angle;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    Lara_DeflectEdgeJump(item, coll);
-    if (item->fall_speed <= 0 || coll->side_mid.floor > 0) {
-        return;
-    }
-
-    if (Lara_LandedBad(item, coll)) {
-        item->goal_anim_state = LS_DEATH;
-    } else {
-        item->goal_anim_state = LS_STOP;
-    }
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
-}
-
-static void M_FallBack(ITEM *const item, COLL_INFO *const coll)
-{
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    lara->move_angle = item->rot.y + DEG_180;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    Lara_DeflectEdgeJump(item, coll);
-
-    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
-        return;
-    }
-
-    if (Lara_LandedBad(item, coll)) {
-        item->goal_anim_state = LS_DEATH;
-    } else {
-        item->goal_anim_state = LS_STOP;
-    }
-
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
 }
 
 static void M_Roll(ITEM *const item, COLL_INFO *const coll)
@@ -771,51 +504,6 @@ static void M_RollContinue(ITEM *const item, COLL_INFO *const coll)
         Lara_ShiftCol(coll);
         item->pos.y += coll->side_mid.floor;
     }
-}
-
-static void M_SwanDive(ITEM *const item, COLL_INFO *const coll)
-{
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    lara->move_angle = item->rot.y;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    Lara_DeflectEdgeJump(item, coll);
-    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
-        return;
-    }
-
-    item->goal_anim_state = LS_STOP;
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
-}
-
-static void M_FastDive(ITEM *const item, COLL_INFO *const coll)
-{
-    LARA_INFO *const lara = Lara_GetLaraInfo();
-    lara->move_angle = item->rot.y;
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = BAD_JUMP_CEILING;
-
-    Lara_GetCollisionInfo(item, coll);
-    Lara_DeflectEdgeJump(item, coll);
-
-    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
-        return;
-    }
-
-    if (item->fall_speed > 133) {
-        item->goal_anim_state = LS_DEATH;
-    } else {
-        item->goal_anim_state = LS_STOP;
-    }
-    item->gravity = false;
-    item->fall_speed = 0;
-    item->pos.y += coll->side_mid.floor;
 }
 
 void Lara_Col_Register(
