+++ src/libtrx/game/lara/col/jump.c
@@ -1,0 +1,321 @@
+#include "config.h"
+#include "game/input.h"
+#include "game/lara.h"
+#include "game/lara/util.h"
+#include "game/sound.h"
+
+static void M_Compress(ITEM *item, COLL_INFO *coll);
+static void M_UpJump(ITEM *item, COLL_INFO *coll);
+static void M_ForwardJump(ITEM *item, COLL_INFO *coll);
+static void M_SideBackJump(ITEM *item, COLL_INFO *coll);
+static void M_FallBack(ITEM *item, COLL_INFO *coll);
+static void M_Reach(ITEM *item, COLL_INFO *coll);
+static void M_SwanDive(ITEM *item, COLL_INFO *coll);
+static void M_FastDive(ITEM *item, COLL_INFO *coll);
+static void M_FastFall(ITEM *item, COLL_INFO *coll);
+
+static void M_Compress(ITEM *const item, COLL_INFO *const coll)
+{
+    item->gravity = false;
+    item->fall_speed = 0;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = NO_BAD_NEG;
+    coll->bad_ceiling = 0;
+
+    Lara_GetCollisionInfo(item, coll);
+
+    if (coll->side_mid.ceiling > -100) {
+        Item_SwitchToAnim(item, LA_STAND_STILL, 0);
+        item->goal_anim_state = LS_STOP;
+        item->current_anim_state = LS_STOP;
+        item->gravity = false;
+        item->speed = 0;
+        item->fall_speed = 0;
+        item->pos = coll->old;
+    }
+
+    if (TR_VERSION >= 2 && coll->side_mid.floor > -STEP_L
+        && coll->side_mid.floor < STEP_L) {
+        item->pos.y += coll->side_mid.floor;
+    }
+}
+
+static void M_UpJump(ITEM *const item, COLL_INFO *const coll)
+{
+#if TR_VERSION == 1
+    const bool enable_lean_jumping = g_Config.gameplay.enable_lean_jumping;
+#else
+    const bool enable_lean_jumping = true;
+#endif
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+    coll->facing = lara->move_angle;
+    if (enable_lean_jumping && item->speed < 0) {
+        coll->facing += DEG_180;
+    }
+
+    Collide_GetCollisionInfo(
+        coll, item->pos.x, item->pos.y, item->pos.z, item->room_num, 870);
+    if (Lara_TestHangJumpUp(item, coll)) {
+        return;
+    }
+
+    Lara_SlideEdgeJump(item, coll);
+    if (enable_lean_jumping) {
+        if (coll->coll_type != COLL_NONE) {
+            item->speed = item->speed > 0 ? 2 : -2;
+        } else if (item->fall_speed < -70) {
+            if (g_Input.forward && item->speed < 5) {
+                item->speed++;
+            } else if (g_Input.back && item->speed > -5) {
+                item->speed -= 2;
+            }
+        }
+    }
+
+    if (item->fall_speed <= 0 || coll->side_mid.floor > 0) {
+        return;
+    }
+
+    if (Lara_LandedBad(item, coll)) {
+        item->goal_anim_state = LS_DEATH;
+    } else {
+        item->goal_anim_state = LS_STOP;
+    }
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+static void M_ForwardJump(ITEM *const item, COLL_INFO *const coll)
+{
+#if TR_VERSION == 1
+    // TODO: TR1's wall bug actually stems from Lara_DeflectEdgeJump, see about
+    // fixing it there.
+    const bool backward_momentum = false;
+    const bool fix_wall_bug = g_Config.gameplay.fix_wall_jump_glitch;
+#else
+    const bool backward_momentum = item->speed < 0;
+    const bool fix_wall_bug = false;
+#endif
+
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    if (backward_momentum) {
+        lara->move_angle = item->rot.y + DEG_180;
+    } else {
+        lara->move_angle = item->rot.y;
+    }
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    Lara_DeflectEdgeJump(item, coll);
+    if (backward_momentum) {
+        lara->move_angle = item->rot.y;
+    }
+
+    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
+        return;
+    }
+
+    if (Lara_LandedBad(item, coll)) {
+        item->goal_anim_state = LS_DEATH;
+    } else if (
+        lara->water_status != LWS_WADE && g_Input.forward && !g_Input.slow) {
+        item->goal_anim_state = LS_RUN;
+    } else {
+        item->goal_anim_state = LS_STOP;
+    }
+
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+    item->speed = 0;
+    if (!fix_wall_bug) {
+        Lara_Animate(item);
+    }
+}
+
+static void M_SideBackJump(ITEM *const item, COLL_INFO *const coll)
+{
+    int32_t angle = 0;
+    switch (item->current_anim_state) {
+    case LS_JUMP_BACK:
+        angle = DEG_180;
+        break;
+    case LS_JUMP_RIGHT:
+        angle = DEG_90;
+        break;
+    case LS_JUMP_LEFT:
+        angle = -DEG_90;
+        break;
+    default:
+        return;
+    }
+
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y + angle;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    Lara_DeflectEdgeJump(item, coll);
+    if (item->fall_speed <= 0 || coll->side_mid.floor > 0) {
+        return;
+    }
+
+    if (Lara_LandedBad(item, coll)) {
+        item->goal_anim_state = LS_DEATH;
+    } else {
+        item->goal_anim_state = LS_STOP;
+    }
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+static void M_FallBack(ITEM *const item, COLL_INFO *const coll)
+{
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y + DEG_180;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    Lara_DeflectEdgeJump(item, coll);
+
+    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
+        return;
+    }
+
+    if (Lara_LandedBad(item, coll)) {
+        item->goal_anim_state = LS_DEATH;
+    } else {
+        item->goal_anim_state = LS_STOP;
+    }
+
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+static void M_Reach(ITEM *const item, COLL_INFO *const coll)
+{
+    item->gravity = true;
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = 0;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    if (Lara_TestHangJump(item, coll)) {
+        return;
+    }
+
+    Lara_SlideEdgeJump(item, coll);
+    if (item->fall_speed <= 0 || coll->side_mid.floor > 0) {
+        return;
+    }
+
+    if (Lara_LandedBad(item, coll)) {
+        item->goal_anim_state = LS_DEATH;
+    } else {
+        item->goal_anim_state = LS_STOP;
+    }
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+static void M_SwanDive(ITEM *const item, COLL_INFO *const coll)
+{
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    Lara_DeflectEdgeJump(item, coll);
+    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
+        return;
+    }
+
+    item->goal_anim_state = LS_STOP;
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+static void M_FastDive(ITEM *const item, COLL_INFO *const coll)
+{
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    Lara_DeflectEdgeJump(item, coll);
+
+    if (coll->side_mid.floor > 0 || item->fall_speed <= 0) {
+        return;
+    }
+
+    if (item->fall_speed > 133) {
+        item->goal_anim_state = LS_DEATH;
+    } else {
+        item->goal_anim_state = LS_STOP;
+    }
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+static void M_FastFall(ITEM *const item, COLL_INFO *const coll)
+{
+    item->gravity = true;
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = BAD_JUMP_CEILING;
+
+    Lara_GetCollisionInfo(item, coll);
+    Lara_SlideEdgeJump(item, coll);
+    if (coll->side_mid.floor > 0) {
+        return;
+    }
+
+    if (Lara_LandedBad(item, coll)) {
+        item->goal_anim_state = LS_DEATH;
+    } else {
+        item->goal_anim_state = LS_STOP;
+        item->current_anim_state = LS_STOP;
+        Item_SwitchToAnim(item, LA_FREEFALL_LAND, 0);
+    }
+
+    Sound_StopEffect(SFX_LARA_FALL);
+    item->gravity = false;
+    item->fall_speed = 0;
+    item->pos.y += coll->side_mid.floor;
+}
+
+// clang-format off
+REGISTER_LARA_COL(LS_COMPRESS,     M_Compress)
+REGISTER_LARA_COL(LS_JUMP_UP,      M_UpJump)
+REGISTER_LARA_COL(LS_JUMP_FORWARD, M_ForwardJump)
+REGISTER_LARA_COL(LS_JUMP_BACK,    M_SideBackJump)
+REGISTER_LARA_COL(LS_JUMP_RIGHT,   M_SideBackJump)
+REGISTER_LARA_COL(LS_JUMP_LEFT,    M_SideBackJump)
+REGISTER_LARA_COL(LS_FALL_BACK,    M_FallBack)
+REGISTER_LARA_COL(LS_REACH,        M_Reach)
+REGISTER_LARA_COL(LS_SWAN_DIVE,    M_SwanDive)
+REGISTER_LARA_COL(LS_FAST_DIVE,    M_FastDive)
+REGISTER_LARA_COL(LS_FAST_FALL,    M_FastFall)
+// clang-format on
