+++ src/libtrx/meson.build
@@ -170,6 +170,7 @@ sources = [
   'game/lara/cheat_keys.c',
   'game/lara/col.c',
   'game/lara/col/climb.c',
+  'game/lara/col/jump.c',
   'game/lara/col/swim.c',
   'game/lara/common.c',
   'game/lara/control.c',
